+++ src/TMC_2209/_TMC_2209_reg.py
@@ -40,6 +40,10 @@ drv_err             = 1<<1
 uv_cp               = 1<<2
 
 #CHOPCONF
+toff0               = 1<<0
+toff1               = 1<<1
+toff2               = 1<<2
+toff3               = 1<<3
 vsense              = 1<<17
 msres0              = 1<<24
 msres1              = 1<<25